Comment Section Component
This is a simple React project which shows a list of comments in a blog. The UI consists of a comment box, inside which at the very top - the avatar, username, timestamp, & their message is present. At the bottom of the box, a couple of buttons are present for Approve & Reject.